@@ -2,28 +2,18 @@ package com.solution.tecno.androidanimations.utils;
 
 import android.app.AlertDialog;
 import android.content.Context;
-import android.os.Build;
 import android.os.Handler;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.TextView;
 
-import androidx.annotation.RequiresApi;
-
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 import com.google.gson.Gson;
 import com.solution.tecno.androidanimations.R;
 
-import java.util.HashMap;
 import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Random;
 
 public class Utils {
@@ -31,7 +21,6 @@ public class Utils {
     private static final String[] ALPHA_NUMERIC_STRING = {"a","A","b","B","c","C","d","D","e","E","f","F","g","G","h","H","i","I","j","J","k","K","l","L","m","M","n","N","o","O","p","P","q","Q","r","R","s","S","t","T","u","U","v","V","w","W","x","X","y","Y","z","Z","0","1","2","3","4","5","6","7","8","9"};
     //Spanish, PeruLocale locale = new Locale("es", "pe"); //Spanish, Peru
     Locale locale = new Locale("es", "pe");
-    private RequestQueue queue;
     private String url ="";
     private Context ctx;
     private Credentials cred;
@@ -91,43 +80,17 @@ public class Utils {
         return eng_date;
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-    public void updateToken(final String token)
+    public FirebaseAuth initFirebaseAuth(){
+        return FirebaseAuth.getInstance();
+    }
+
+    public String objectToJson(Object obj){
+        return new Gson().toJson(obj);
+    }
+
+    public DatabaseReference getDatabaseReference(String database)
     {
-        cred = new Credentials(ctx);
-        try{
-            String url=ctx.getApplicationContext().getString(R.string.base_url)+"updateFCM.php";
-            Log.i("updateToken_url",url);
-            queue = Volley.newRequestQueue(ctx);
-
-            // Request a string response from the provided URL.
-            StringRequest stringRequest = new StringRequest(Request.Method.POST, url,
-                    new Response.Listener<String>() {
-                        @Override
-                        public void onResponse(String response) {
-                            System.out.println("updateToken_response: " + response);
-                        }
-                    }, new Response.ErrorListener() {
-                @Override
-                public void onErrorResponse(VolleyError error) {
-                    System.out.println("updateToken_error: " + error.getMessage());
-                }
-            }){
-                @Override
-                protected Map<String, String> getParams()
-                {
-                    Map<String, String> params = new HashMap<>();
-                    params.put("id",cred.getUserId());
-                    params.put("fcm",token);
-                    return params;
-                }
-            };
-            // Add the request to the RequestQueue.
-            queue.add(stringRequest);
-        }catch(Exception e){
-            Log.e("updateToken_error", Objects.requireNonNull(e.getMessage()));
-        }
-
-
+        FirebaseDatabase databaseReference = FirebaseDatabase.getInstance(ctx.getString(R.string.firebase_database));
+        return databaseReference.getReference(database);
     }
 }
